feat(search): updated search definitions to please Kotlin client

+++ projects/api/src/contracts/search/_internal/response/searchPersonResponseSchema.ts
@@ -1,8 +1,8 @@
-import { z } from '../../../_internal/z.ts';
+import { int64, z } from '../../../_internal/z.ts';
 import { peopleSummaryResponseSchema } from '../../../people/_internal/response/peopleSummaryResponseSchema.ts';
 
 export const searchPersonResponseSchema = z.object({
-  score: z.number().int(),
+  score: int64(z.number().int()),
   type: z.literal('person'),
   person: peopleSummaryResponseSchema,
 });
